Make point class use ints

+++ helpers/point/point.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 
 @dataclass(frozen=True)
 class Point:
-    x: float
-    y: float
+    x: int
+    y: int
